fix: correct encoding in tests
