+++ crates/core/src/lib.rs
@@ -12,8 +12,7 @@ pub mod types;
 
 // Re-export commonly used items
 pub use encoding::{
-    TURETZKY_UPPER_GAS_LIMIT, decode_state_updates_tuple, encode_state_updates_to_abi,
-    encode_state_updates_to_sol,
+    TURETZKY_UPPER_GAS_LIMIT, encode_state_updates_to_abi, encode_state_updates_to_sol,
 };
 pub use heuristic::{
     BASE_TX_COST, LOG_BASE_COST, LOG_DATA_COST_PER_BYTE, LOG_TOPIC_COST, TraceOperations,
